# 🚀 perf: optimize client performance and update benchmarks